requirements.txt
Flask
firebase-admin
Werkzeug
openai
python-dotenv
requests
